requirements.txt
python-dotenv==1.0.0

langchain
langchain-core

langchain-huggingface
sentence-transformers
fastembed

qdrant-client
langchain-qdrant